requirements.txt
streamlit==1.20.0
azure-cognitiveservices-speech==1.26.0
pytube==12.1.2
moviepy==1.0.3
